Made great progress on multiplayer code, figuring out how best to represent packets.

+++ Server/src/main/kotlin/com/example/PlayerConnection.kt
@@ -11,7 +11,7 @@ class PlayerConnection(val session: DefaultWebSocketSession) {
 
     val id = ID.getAndIncrement()
 
-    var player = Player(id.toLong(), 0.0, 0.0)
+    var player = Player(id, 0.0, 0.0)
 
     override fun toString(): String {
         return "PlayerConnection(id=$id)"
